azure_ai_document_intelligence/requirements.txt
azure-identity
azure-mgmt-cognitiveservices